% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{abundancePlot}
\alias{abundancePlot}
\title{Boxplot with abundance of relevant species}
\usage{
abundancePlot(ps, x = "Species", fill = "Genus")
}
\arguments{
\item{ps}{(Required) An class S4 class phyloseq object.}

\item{x}{(Optional, default="Species") A character string. Representing a taxonomic level.}

\item{fill}{(Optional, default="Genus") A character string. Representing a taxonomic level.}
}
\value{
A DEseq2 results table with significant expression level values.
}
\description{
This function make boxplots of filtered phyloseq objects.
}
\examples{
data(ps_16S)
ps_16S <- subset_samples(ps_16S, SORTED_genus == "Synchaeta")
ps_16S <- relevantSpecies(ps_16S, method = "vst", cutoff = 0.75, design = ~ ID_friendly)
plot <- abundancePlot(ps_16S, x = "Species", fill = "Genus")

}
