% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{barChart}
\alias{barChart}
\title{barCharts out of a summaryTable using ggplot2}
\usage{
barChart(sumtable, x = "Genus", fill = "Order", errorbars = TRUE, ...)
}
\arguments{
\item{sumtable}{(Required) A dataframe (preferably the output of metaphylo::sumTable function)}

\item{x}{(Optional, default="Genus") Character string referring to the main explainatory variable.}

\item{fill}{(optional, default="Order") A string reffereng to any aditional explainatory variable for filling.}

\item{errorbars}{(Optional, default=TRUE) Boolean operator. If True, standard error is plotted as errorbars.}

\item{...}{Any aditional variables passed on to geom_bar(...)}
}
\value{
A ggplot object. The aim of the data frame is to use for plotting in ggplot (or the wrapper metaphylo::barChart)
}
\description{
This function takes a phyloseq as unput (or possibly a tidy data-frame) and transforms it into a summary table.
}
\examples{
data(ps_18S)
ps_18S \%>\%
 subset_samples(SORTED_genus=="Synchaeta") \%>\%
 subset_samples(MONTH=="aug") \%>\%
 subset_taxa(Class!="Rotifera") \%>\%
 transform_sample_counts(function(x) (x/sum(x))*100) \%>\%
 filter_taxa(function(x) sum(x>1)>=((length(x)*0.5)),TRUE) \%>\%
 sumTable(x = "Species", variables = c("Class", "Order")) \%>\%
 barChart(x = "Family", fill = "Class")
}
