% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{colorWeb}
\alias{colorWeb}
\title{LARGE WRAPPER: Pipartite plot of FOC and RRA data}
\usage{
colorWeb(
  dualmat,
  foc.color = c("gray35", "gray40", "gray45", "gray50", "gray90"),
  method = "normal",
  text.rot = -90,
  labsize = 0.5,
  bor.col.interaction = NULL,
  ...
)
}
\arguments{
\item{dualmat}{(Required) The output of summarizedPhyloseq (i.e. a list of two matrixes.)}

\item{foc.color}{(Optional, c("gray35", "gray40", "gray45", "gray50", "gray90")).}

\item{...}{(Optional). Refer to bipartite::plotweb options.}

\item{col.order}{(Optional) Vector of strings, representing column order.}

\item{row.order}{(Optional) Vector of strings, representing row order.}
}
\value{
A bipartite::plotweb plot.
}
\description{
Produces bipartite plot out of summarized phyloseqobject (output of summarizedPhyloseq function)
Thickness of interaction represents Relative read abundance. Color of interaction represents Frequency of occurance.
}
\examples{
data(ps_18S)
ps_18S <- subset_samples(ps_18S, SORTED_genus == "Synchaeta") \%>\%
tax_glom("Family") \%>\%
summarizedPhyloseq(variable="MONTH")

}
