% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{CompleteTaxonomy}
\alias{CompleteTaxonomy}
\title{Fills unassigned taxonomic levels with useful information.}
\usage{
CompleteTaxonomy(tax, string = "_x")
}
\arguments{
\item{tax}{(Required) A list of two matrices. As returned from DADA2::assignTaxonomy(seqtab, database, outputBootstraps =  TRUE).}

\item{string}{(Optional, default="_x") A string. Will be attached at the end of the name comming from the higher taxonomic level.}
}
\value{
A list of two matrices (modified tax object).
}
\description{
This function assigns the higher taxonomic ranks to unassigned taxoomic levels, using the output of DADA2::assignTaxonomy.
}
\examples{
tax1 <- CompleteTaxonomy(tax)
tax2 <- CompleteTaxonomy(tax, string="X")
}
