% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{filterTaxonomy}
\alias{filterTaxonomy}
\title{Filters DADA2 taxonomy table for low bootstrap values.}
\usage{
filterTaxonomy(tax, minboot = 75)
}
\arguments{
\item{tax}{(Required) A list of two matrices. As returned from DADA2::assignTaxonomy(seqtab, database, outputBootstraps =  TRUE).}

\item{minboot}{(Optional, default=75) A Number between 0-100. Minimal accepted bootstrap value.}
}
\value{
A list of two matrices (modified tax object).
}
\description{
This function takes the output taxonomy table from the DADA2::assignTaxonomy and filters it based on bootstrap value.
}
\examples{
data("tax_16S")
tax_16S <- filterTaxonomy(tax_16S, minboot = 75)

}
