% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{foodweb}
\alias{foodweb}
\title{Plot a foodweb with statistical DESeq2 analysi results, using bipartites plotweb.}
\usage{
foodweb(ps, design, group, sigcol = "Red", nonsigcol = "Green",
  medsigcol = "Blue", ...)
}
\arguments{
\item{ps}{(Required) An S4 class phyloseq object.}

\item{design}{(Required) A ~ variable character string. Representing a column in the sample dataset.}

\item{group}{(Required) Character string.}

\item{sigcol}{(Optional, default="Red") Character string. R-base color to represent significans.}

\item{nonsigcol}{(Optional, default="Green") Character string. R-base color to represent significans.}

\item{medsigcol}{(Optional, default="Blue") Character string. R-base color to represent significans.}

\item{...}{(Opltional) See bipartite::plotweb for more options.}
}
\value{
A bipartite::plotweb graph.
}
\description{
Plot a foodweb with statistical DESeq2 analysi results, using bipartites plotweb.
}
\examples{
ps <- ps_18S \%>\%
subset_samples(MONTH=="aug") \%>\% # Synchaeta is selected
subset_samples(SORTED_type=="Rotifer") \%>\% # Synchaeta from may is removed.
subset_taxa(Class!="Rotifera") \%>\%
relevantSpecies(method = "vst", cutoff=0.3, design = ~SORTED_genus) \%>\%
foodweb(design= ~SORTED_genus, group= "SORTED_genus")
}
