% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{GetSample}
\alias{GetSample}
\title{Extract Sample metadata table of a phyloseq object}
\usage{
GetSample(ps)
}
\arguments{
\item{PHYLOSEQ}{(Required) An class S4 class phyloseq object.}
}
\value{
A data frame
}
\description{
Extract Sample metadata table of a phyloseq object
}
