% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{getSignificant}
\alias{getSignificant}
\title{Filters Phyloseq object by taxa with significantly different abundance}
\usage{
getSignificant(ps, design, number = 50, alpha = 0.01, vst = FALSE)
}
\arguments{
\item{ps}{(Required) An S4 class phyloseq object.}

\item{design}{(Required) A ~ variable character string. Representing a column in the sample dataset.}

\item{number}{(Optional, default=50) An integer. Biggest number of taxa to keep in the filtered dataset.}

\item{alpha}{(Optional, default=0.01) A value. Threshold for p-value.}

\item{vst}{(Optional, default=FALSE) Boolean operator. If TRUE, the dataset gets variance stabalising transformation.}
}
\value{
An S4 class phyloseq object, filtered
}
\description{
Filters Phyloseq object by taxa with significantly different abundance
}
\examples{
data(ps_16S)
ps_16S <- subset_samples(ps_16S, SORTED_genus == "Synchaeta")
ps_16S_sig <- getSignificant(ps_16S, ~MONTH)
}
