% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{GetTax}
\alias{GetTax}
\title{Extract Taxonomy table of a phyloseq object}
\usage{
GetTax(ps)
}
\arguments{
\item{PHYLOSEQ}{(Required) An class S4 class phyloseq object.}
}
\value{
A matrix.
}
\description{
Extract Taxonomy table of a phyloseq object
}
