% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicheoverlap.R
\name{nicheOverlap}
\alias{nicheOverlap}
\title{Schoeners Index of Nieche Overlap}
\usage{
nicheOverlap(mat, species1, species2)
}
\arguments{
\item{mat}{(Required) A food web matrix, with consurs as columns and prey as rows. Relative proportion.}

\item{species1}{(Required) Character string, Colname of Species 1}

\item{species2}{(Required) Character string, Colname of Species 2}
}
\value{
Double, percentage of Diet Overlap.
}
\description{
Calculates the percentage overlap of diet between two species in a food web matrix.
}
