% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicheoverlap.R
\name{nicheOverlapPerSample}
\alias{nicheOverlapPerSample}
\title{Schoeners Index of Nieche Overlap, Per Sample.}
\usage{
nicheOverlapPerSample(ps, sample_group = "SORTED_genus")
}
\arguments{
\item{ps}{(Required) A S4 type Phyloseq object. OBS currently tax-glomed to Family level.}

\item{sample_group}{(Optional, defaut="SORTED_genus). A character string, grouping variable.}
}
\value{
A dataframe, with
}
\description{
Calculates the percentage overlap of diet between ALL possible pairs of samples in a phyloseq object,
}
