% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicheoverlap.R
\name{pairwiseNicheOverlap}
\alias{pairwiseNicheOverlap}
\title{Schoeners Pairwise Index of Nieche Overlap}
\usage{
pairwiseNicheOverlap(mat)
}
\arguments{
\item{mat}{(Required) A food web matrix, with consurs as columns and prey as rows. Relative proportion.}
}
\value{
A matrix, of type distance matrix, percentage of Diet Overlap.
}
\description{
Calculates the percentage overlap of diet between ALL pairs of species in a food web matrix.
}
