% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{pColors}
\alias{pColors}
\title{Not exported. Fit colors to matrix.}
\usage{
pColors(
  psmatrix,
  pvector,
  sigcol = "Red",
  nonsigcol = "Green",
  medsigcol = "Blue"
)
}
\arguments{
\item{psmatrix}{(Required) A matrix (see phyloseqToMatrix).}

\item{pvector}{(Required) A named vector with of p-values per taxa (see pVector)}

\item{sigcol}{(Optional, default="Red") Character string. R-base color to represent significans.}

\item{nonsigcol}{(Optional, default="Green") Character string. R-base color to represent significans.}

\item{medsigcol}{(Optional, default="Blue") Character string. R-base color to represent significans.}
}
\value{
A vector with colrs.
}
\description{
Not exported. Fit colors to matrix.
}
