% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{phyloseqBarChart}
\alias{phyloseqBarChart}
\title{Plot BarCharts from phyloseq object}
\usage{
phyloseqBarChart(ps, x = "Genus", variables = c("Order"),
  errorbars = TRUE, ...)
}
\arguments{
\item{ps}{(Required) Phyloseq object}

\item{x}{(Optional, default="Genus") Character string referring to the main explainatory variable.}

\item{variables}{(optional, default=c("Order")) A vector of strings reffereng to any aditional explainatory variables. First string will be used for "fill".}

\item{errorbars}{(Optional, default=TRUE) Boolean operator. If True, standard error is plotted as errorbars.}

\item{...}{Any aditional variables passed on to geom_bar(...)}
}
\value{
A ggplot object.
}
\description{
This is a wrapper function around the two functions sumTable and barChart for phyloseq objects.
}
\examples{
data(ps_18S)
ps_18S \%>\%
 subset_samples(SORTED_genus=="Synchaeta") \%>\%
 subset_samples(MONTH=="aug") \%>\%
 subset_taxa(Class!="Rotifera") \%>\%
 transform_sample_counts(function(x) (x/sum(x))*100) \%>\%
 filter_taxa(function(x) sum(x>1)>=((length(x)*0.5)),TRUE) \%>\%
 phyloseqBarChart(x="Genus", variables=c("Order"))
}
