% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{PhyloseqToCSV}
\alias{PhyloseqToCSV}
\title{Writes a phyloseq object as standard OTU-like table}
\usage{
PhyloseqToCSV(PHYLOSEQ, HEADER, FILENAME = "./ASV.csv")
}
\arguments{
\item{PHYLOSEQ}{(Required) An class S4 class phyloseq object}

\item{HEADER}{(Required) A character string. Representing a column in the sample dataset.}

\item{FILENAME}{(Optional, defalt="./ASV.csv") Character vector, filename.}
}
\value{
A S4 class Phyloseq object.
}
\description{
This function takes a full phyloseq object and merges the count matrix with the taxonomy table. It also provides the option to give sample names.
}
\examples{
ps <- PhyloseqToCSV(ps, "SORTED_type", "./phyloseq.csv")
}
