% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{phyloseqToMatrix}
\alias{phyloseqToMatrix}
\title{Writes a phyloseq object as matrix, specifying rownames as prey and colnames as species.}
\usage{
phyloseqToMatrix(ps, colname, rowname = "Species")
}
\arguments{
\item{ps}{(Required) An class S4 class phyloseq object}

\item{colname}{(Required) A character string. Representing a column in the sample dataset.}

\item{rowname}{(Optional, defalt="Species") A character string. Representing a column in the sample dataset.}
}
\value{
A S4 class Phyloseq object.
}
\description{
This function takes a full phyloseq object and merges the count matrix with the taxonomy table. It also provides the option to give sample names.
}
\examples{
data(ps_18S)
phyloseqToMatrix(ps_18S, colnames="ID_friendly")
}
