% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{prevalentSpecies}
\alias{prevalentSpecies}
\title{Filtering of prevalent species.}
\usage{
prevalentSpecies(ps, prevalence = 0.75, minpercent = 0.1)
}
\arguments{
\item{ps}{(Required) A class S4 class phyloseq object.}

\item{prevalence}{(Optional, default=0.75) A numeric value between 0:1. Least fraction of required presence.}

\item{minpercent}{(Optional, default=0.1) Numeric value between 0:100. A percentage of total library size. Lowest number to count as presence.}
}
\value{
A class S4 class phyloseq object.
}
\description{
This function filters away taxa only present in few samples.
}
\examples{
data(ps_16S)
ps_16S <- subset_samples(ps_16S, SORTED_genus == "Synchaeta")
ps_16S <- prevalentSpecies(ps_16S, prevalence=0.75, minpercent=1)
}
