% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{pVector}
\alias{pVector}
\title{DESeq2 analysis: p.adj named vector.}
\usage{
pVector(ps, design)
}
\arguments{
\item{ps}{(Required) An S4 class phyloseq object.}

\item{design}{(Required) A ~ variable character string. Representing a column in the sample dataset.}
}
\value{
A Named vector of p.adj values per species.
}
\description{
DESeq2 analysis: p.adj named vector.
}
\examples{
data(ps_18S)
}
