% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{select_samples}
\alias{select_samples}
\title{Subset phyloseq object based on sample data}
\usage{
select_samples(physeq, variable, value)
}
\arguments{
\item{physeq}{(Required) Phyloseq object}

\item{variable}{(Required) Character string referring to variable name in Sample data.}

\item{value}{(Required) Character string, or vector of character string.}
}
\value{
Philtered phyloseq
}
\description{
Instead of using phyloseq::subset_samples.
}
\examples{
data(ps_18S)
ps_18S \%>\%
 select_samples("SORTED_genus","Keratella")
ps_18S \%>\%
 select_samples("SORTED_genus",c("Keratella", "Synchaeta"))
}
