% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seqtab_16S}
\alias{seqtab_16S}
\title{Read qount matrix for 16S gene}
\format{A matrix with raw read counts as achieved by dada2 R package pipeline, with sample names as column names and exact sequence as row name.}
\source{
Exact analysis code to generate this data \url{https://github.com/andreasnovotny/DadaSlurm}
}
\usage{
seqtab_16S
}
\description{
Plankton foodweb metabarcoding project, Stockholm University 2017-2018
Data generated on Illumina Miseq platform.
}
\keyword{datasets}
