% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{sortByWater}
\alias{sortByWater}
\title{Sort a phyloseq object by presence of prevalent taxa in another phyloseq object}
\usage{
sortByWater(ps_1, ps_2, minpercent = 1)
}
\arguments{
\item{ps_1}{(Required) A phyloseq object to be filtered.}

\item{ps_2}{(Required) A phyloseq object to filter by.}

\item{minpercent}{(Optional, default=1) Character string. R-base color to represent significans.}
}
\value{
A filtered phyloseq object.
}
\description{
A wrapper function for phyloseq::prune_taxa
}
\examples{
data(ps_18S)

}
