% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{summarizedPhyloseq}
\alias{summarizedPhyloseq}
\title{LARGE WRAPPER: Summarizing RRA and FOC for phyloseq objet}
\usage{
summarizedPhyloseq(
  ps = synchaeta_18S,
  variable = MONTH,
  rowname = Family,
  prevalence = 0.6
)
}
\arguments{
\item{ps}{(Required) A class S4 class phyloseq object.}

\item{variable}{(Required) Variable name in Sample data. Subseting variable.}

\item{rowname}{(Required) Only evaluated if PS_output=FALSE. This has to be the taxonomic rank with the highest available resolution. (Refer to phyloseq::tax_glom)}

\item{prevalence}{(Optional, default=0.6) A numeric value between 0:1. Least fraction of required presence.}
}
\value{
A list of two phyloseq objects, OR A list of two matrices. (See PS_output)
}
\description{
This function filters away taxa only present in few samples.
Constructs Relative read count tables and frequency of occurance tables
}
\examples{
data(ps_18S)
ps_18S <- subset_samples(ps_18S, SORTED_genus == "Synchaeta") \%>\%
tax_glom("Family") \%>\%
summarizedPhyloseq(variable="MONTH")

}
