% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{sumTable}
\alias{sumTable}
\title{Construct Summary table for barplots}
\usage{
sumTable(
  data,
  x = "Genus",
  y = "Abundance",
  variables = c("Family", "Class", "Order")
)
}
\arguments{
\item{data}{(Required) A phyloseq object or a data frame.}

\item{x}{(Optional, default="Genus") Character string referring to the main explainatory variable.}

\item{y}{(Optional, default="Abundance") Character string reffering to the dependent variable. For phyloseq objects, this is always "Abundance".}

\item{variables}{(optional, default=c("Family", "Class", "Order")) A vector of strings reffereng to any aditional explainatory variable for plotting/colouring/splitting/grouping ect.}
}
\value{
A formated data.frame. The aim of the data frame is to use for plotting in ggplot (or the wrapper metaphylo::barChart)
}
\description{
This function takes a phyloseq as unput (or possibly a tidy data-frame) and transforms it into a summary table.
}
\examples{
data(ps_18S)
ps_18S \%>\%
 subset_samples(SORTED_genus=="Synchaeta") \%>\%
 subset_samples(MONTH=="aug") \%>\%
 subset_taxa(Class!="Rotifera") \%>\%
 transform_sample_counts(function(x) (x/sum(x))*100) \%>\%
 filter_taxa(function(x) sum(x>1)>=((length(x)*0.5)),TRUE) \%>\%
 sumTable(x = "Species", variables = c("Class", "Order"))
}
