% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tax_16S}
\alias{tax_16S}
\title{Taxonomic information for 16S gene}
\format{A list containing to matrixes, as generated by the dada2 R package pipeline:#' \describe{
\item{$tax}{Taxonomic affiliation for each exact sequence variant}
\item{$boot}{Bootstrap value of each taxonomic affiliation}
}}
\source{
Exact analysis code to generate this data \url{https://github.com/andreasnovotny/DadaSlurm}
}
\usage{
tax_16S
}
\description{
Plankton foodweb metabarcoding project, Stockholm University 2017-2018
Data generated on Illumina Miseq platform.
}
\keyword{datasets}
