% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{completeTaxonomy}
\alias{completeTaxonomy}
\title{Fills unassigned taxonomic levels with useful information.}
\usage{
completeTaxonomy(tax, string = "_x")
}
\arguments{
\item{tax}{(Required) A list of two matrices. As returned from DADA2::assignTaxonomy(seqtab, database, outputBootstraps =  TRUE).}

\item{string}{(Optional, default="_x") A string. Will be attached at the end of the name comming from the higher taxonomic level.}
}
\value{
A list of two matrices (modified tax object).
}
\description{
This function assigns the higher taxonomic ranks to unassigned taxoomic levels, using the output of DADA2::assignTaxonomy.
}
\examples{
data("tax_16S")
tax_16S <- filterTaxonomy(tax_16S, minboot = 75)
tax_16S <- completeTaxonomy(tax_16S, string = "_x")

}
