% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{dada2ToPhyloseq}
\alias{dada2ToPhyloseq}
\title{Create phyloseq object from DADA2 output.}
\usage{
dada2ToPhyloseq(seqtab, tax, metadata)
}
\arguments{
\item{seqtab}{(Required) A sequence table matrix as returned from dada2::makeSequenceTable()}

\item{tax}{(Required) A list of two matrices. As returned from DADA2::assignTaxonomy(seqtab, database, outputBootstraps =  TRUE).}

\item{metadata}{(Required) A data frame.}
}
\value{
A S4 class Phyloseq object.
}
\description{
This function takes the output from DADA2 package and returns a complete phyloseq object.
}
\examples{
data("metadata_16S")
data("seqtab_16S")
data("tax_16S")
ps_16S <- dada2ToPhyloseq(seqtab_16S, tax_16S, metadata_16S)

}
