% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{deseqTable}
\alias{deseqTable}
\title{Makes a DESeq2 analysis of a phyloseq object}
\usage{
deseqTable(ps, design, speciesAsNames = TRUE)
}
\arguments{
\item{speciesAsNames}{(Optional) Boolean (Default = TRUE). If true, species names will be the tables row name.}

\item{PHYLOSEQ}{(Required) An class S4 class phyloseq object}

\item{VARIABLE}{(Required) A ~ variable character string. Representing a column in the sample dataset.}
}
\value{
A DEseq2 results table with significant expression level values.
}
\description{
This function makes a DESeq2 stastistical analysis and writes the results to a redable table.
}
\examples{
ps <- DEseqTable(ps, ~ SORTED_type)
}
