% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{foodWeb}
\alias{foodWeb}
\title{A wrapper function for bipartite::plotweb}
\usage{
foodWeb(psmatrix, pvector, col.high = "blue4", bor.col.high = "blue4",
  bor.col.low = "white", col.interaction = c("darkgray", "dimgrey"),
  bor.col.interaction = c("dimgrey", "darkgrey"), sigcol = "Red",
  nonsigcol = "Green", medsigcol = "Blue", ...)
}
\arguments{
\item{psmatrix}{(Required) A matrix (see phyloseqToMatrix).}

\item{pvector}{(Required) A named vector with of p-values per taxa (see pVector)}

\item{sigcol}{(Optional) Character string. R-base color to represent significans.}

\item{nonsigcol}{(Optional, default="Green") Character string. R-base color to represent significans.}

\item{medsigcol}{(Optional, default="Blue") Character string. R-base color to represent significans.}

\item{...}{(Opltional) See bipartite::plotweb for more options.}
}
\value{
A bipartite::plotweb graph.
}
\description{
A wrapper function for bipartite::plotweb
}
\examples{
data(ps_18S)

}
