% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{getVst}
\alias{getVst}
\title{Variance stabilizing transformation of phyloseq object using DESeq2}
\usage{
getVst(ps, design)
}
\arguments{
\item{ps}{(Required) An S4 class phyloseq object.}

\item{design}{(Required) A ~ variable character string. Representing a column in the sample dataset.}
}
\value{
An S4 class phyloseq object, with transformed counts.
}
\description{
Variance stabilizing transformation of phyloseq object using DESeq2
}
\examples{
ps_18S_1 <- getVST(ps_18S, ~MONTH+SORTED_genus)
}
