% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{relevantSpecies}
\alias{relevantSpecies}
\title{Filtering of relevant species}
\usage{
relevantSpecies(ps, method = "logRaw", cutoff = 0.75, mincount = 10,
  design)
}
\arguments{
\item{ps}{(Required) A class S4 class phyloseq object.}

\item{method}{(Optional, default="logRaw") "logRaw" (log transformation) or "vst" (variance stabilising DESeq2 transformation)}

\item{cutoff}{(Optional, default=0.8) A numeric value between 0:1.}

\item{mincount}{(Optional, default=10) An integer, bigger than 0. Abundance value as been counted as presence. Only effective if method="logRaw".}

\item{dsign}{(Required if method="rsv") A ~ variable character string. Representing a column in the sample dataset.}
}
\value{
A class S4 class phyloseq object.
}
\description{
This function filters away taxa only present in few samples.
}
\examples{
data(ps_16S)
ps_16S <- subset_samples(ps_16S, SORTED_genus == "Synchaeta")
ps_16S <- relevantSpecies(ps_16S, method = "vst", cutoff = 0.75, design = ~ ID_friendly)
}
